+++ modules/meeting/app/components/_index.sass
@@ -7,4 +7,4 @@
 @import "./meeting_agenda_items/outcomes/outcome_component.sass"
 @import "./meetings/header_infoline_component.sass"
 @import "./meetings/participants/add_user_form_component.sass"
-@import "./meetings/presentation_component.sass"
+@import "./meetings/presentation_mode/show_component.sass"
