--- modules/meeting/app/components/meetings/presentation_component.html.erb
@@ -1,111 +1,0 @@
-<%= component_wrapper(
-      class: "op-meeting-presentation",
-      data: {
-        controller: "meetings--presentation poll-for-changes zen-mode",
-        zen_mode_autostart_value: @initial,
-        poll_for_changes_continuous_value: true,
-        poll_for_changes_url_value: check_for_updates_project_meeting_presentation_path(
-          @meeting.project,
-          @meeting,
-          meeting_agenda_item_id: current_item.id,
-          started_at: started_at_param
-        ),
-        poll_for_changes_interval_value: check_for_updates_interval,
-      }
-    ) do %>
-  <%=
-    render Primer::OpenProject::PageHeader.new(classes: "op-meeting-presentation-header") do |header|
-      header.with_breadcrumbs([])
-      header.with_title_content(@meeting.title)
-
-      header.with_action_button(
-        tag: :a,
-        scheme: :invisible,
-        mobile_icon: :x,
-        mobile_label: t(:"meeting.presentation_mode.exit"),
-        href: meeting_path(@meeting),
-        aria: { label: t(:"meeting.presentation_mode.exit") },
-        title: t(:"meeting.presentation_mode.exit"),
-      ) do |button|
-        button.with_leading_visual_icon(icon: :x)
-        t(:"meeting.presentation_mode.exit")
-      end
-    end
-  %>
-  <%= content_tag :div,
-                  id: "op-meeting-presentation-content",
-                  class: "op-meeting-presentation--content",
-                  data: {
-                    controller: "meetings--submit",
-                    poll_for_changes_target: "reference",
-                    reference_value: current_item.updated_at.iso8601
-                  } do %>
-    <div class="op-meeting-presentation--header">
-      <div class="op-meeting-presentation--header-title">
-        <%= render(Primer::Beta::Text.new(color: :subtle)) { MeetingSection.model_name.human } %>:
-        <%= render(Primer::Beta::Text.new(font_weight: :bold, color: :subtle)) { current_section.title } %>
-      </div>
-    </div>
-
-    <%= render(border_box_container) do |container| %>
-      <%= render MeetingAgendaItems::ItemComponent.new(
-        container:,
-        meeting_agenda_item: current_item,
-        presentation_mode: true
-      ) %>
-    <% end %>
-  <% end %>
-
-  <div class="op-meeting-presentation--footer">
-    <div class="op-meeting-presentation--footer-left">
-      <%= render(Primer::Beta::Button.new(
-        tag: :a,
-        scheme: :link,
-        href: project_meeting_presentation_path(@project, @meeting, current_id: @current_item, action_type: "previous", started_at: started_at_param),
-        data: { "meetings--presentation-target": "previousButton" },
-        disabled: !has_previous?
-      )) do |button| %>
-        <% button.with_leading_visual_icon(icon: :"chevron-left") %>
-        <%= t('meeting.presentation_mode.previous') %>
-      <% end %>
-      <% if previous_item %>
-        <div class="op-meeting-presentation--footer-agenda-item">
-          <% if previous_item.meeting_section != current_section %>
-            <%= render(Primer::Beta::Text.new(font_weight: :bold, color: :subtle)) { previous_item.meeting_section.title } %>:
-          <% end %>
-          <%= render(Primer::Beta::Text.new(color: :subtle)) { previous_item.display_title } %>
-        </div>
-      <% end %>
-    </div>
-
-    <div class="op-meeting-presentation--progress">
-      <div class="op-meeting-presentation--progress-text">
-        <%= progress_text %>
-      </div>
-      <div class="op-meeting-presentation--timer">
-        <%= running_time %>
-      </div>
-    </div>
-
-    <div class="op-meeting-presentation--footer-right">
-      <%= render(Primer::Beta::Button.new(
-        tag: :a,
-        scheme: :link,
-        href: project_meeting_presentation_path(@project, @meeting, current_id: @current_item, action_type: "next", started_at: started_at_param),
-        data: { "meetings--presentation-target": "nextButton" },
-        disabled: !has_next?
-      )) do |button| %>
-        <%= t('meeting.presentation_mode.next') %>
-        <% button.with_trailing_visual_icon(icon: :"chevron-right") %>
-      <% end %>
-      <div class="op-meeting-presentation--footer-agenda-item">
-        <% if next_item %>
-          <% if next_item.meeting_section != current_section %>
-            <%= render(Primer::Beta::Text.new(font_weight: :bold, color: :subtle)) { next_item.meeting_section.title } %>:
-          <% end %>
-          <%= render(Primer::Beta::Text.new(color: :subtle)) { next_item.display_title } %>
-        <% end %>
-      </div>
-    </div>
-  </div>
-<% end %>
