+++ modules/meeting/app/components/meetings/presentation_mode/footer_component.html.erb
@@ -1,0 +1,53 @@
+<div class="op-meeting-presentation--footer">
+  <div class="op-meeting-presentation--footer-left">
+    <%= render(Primer::Beta::Button.new(
+      tag: :a,
+      scheme: :link,
+      href: project_meeting_presentation_path(@project, @meeting, current_id: @current_item.id, action_type: "previous", started_at: @started_at),
+      data: { "meetings--presentation-target": "previousButton" },
+      disabled: !has_previous?
+    )) do |button| %>
+      <% button.with_leading_visual_icon(icon: :"chevron-left") %>
+      <%= t('meeting.presentation_mode.previous') %>
+    <% end %>
+    <% if previous_item %>
+      <div class="op-meeting-presentation--footer-agenda-item">
+        <% if previous_item.meeting_section != current_section %>
+          <%= render(Primer::Beta::Text.new(font_weight: :bold, color: :subtle)) { previous_item.meeting_section.title } %>:
+        <% end %>
+        <%= render(Primer::Beta::Text.new(color: :subtle)) { previous_item.display_title } %>
+      </div>
+    <% end %>
+  </div>
+
+  <div class="op-meeting-presentation--progress">
+    <div class="op-meeting-presentation--progress-text">
+      <%= progress_text %>
+    </div>
+    <div class="op-meeting-presentation--timer">
+      <%= running_time %>
+    </div>
+  </div>
+
+  <div class="op-meeting-presentation--footer-right">
+    <%= render(Primer::Beta::Button.new(
+      tag: :a,
+      scheme: :link,
+      href: project_meeting_presentation_path(@project, @meeting, current_id: @current_item.id, action_type: "next", started_at: @started_at),
+      data: { "meetings--presentation-target": "nextButton" },
+      disabled: !has_next?
+    )) do |button| %>
+      <%= t('meeting.presentation_mode.next') %>
+      <% button.with_trailing_visual_icon(icon: :"chevron-right") %>
+    <% end %>
+    <div class="op-meeting-presentation--footer-agenda-item">
+      <% if next_item %>
+        <% if next_item.meeting_section != current_section %>
+          <%= render(Primer::Beta::Text.new(font_weight: :bold, color: :subtle)) { next_item.meeting_section.title } %>:
+        <% end %>
+        <%= render(Primer::Beta::Text.new(color: :subtle)) { next_item.display_title } %>
+      <% end %>
+    </div>
+  </div>
+</div>
+
