+++ modules/meeting/app/components/meetings/presentation_mode/header_component.html.erb
@@ -1,0 +1,20 @@
+<%=
+  render Primer::OpenProject::PageHeader.new(classes: "op-meeting-presentation-header") do |header|
+    header.with_breadcrumbs([])
+    header.with_title_content(@meeting.title)
+
+    header.with_action_button(
+      tag: :a,
+      scheme: :invisible,
+      mobile_icon: :x,
+      mobile_label: t(:"meeting.presentation_mode.exit"),
+      href: meeting_path(@meeting),
+      aria: { label: t(:"meeting.presentation_mode.exit") },
+      title: t(:"meeting.presentation_mode.exit"),
+    ) do |button|
+      button.with_leading_visual_icon(icon: :x)
+      t(:"meeting.presentation_mode.exit")
+    end
+  end
+%>
+
