+++ modules/meeting/app/components/meetings/presentation_mode/show_component.sass
@@ -79,3 +79,4 @@
     height: 100vh
     width: 100vw
     padding: 50px
+
