+++ modules/meeting/app/views/meeting_presentation/show.html.erb
@@ -1,7 +1,7 @@
 <% html_title "#{t(:label_meeting)}: #{@meeting.title} - #{t('meeting.presentation_mode.title')}" %>
 
 <%=
-  render Meetings::PresentationComponent.new(
+  render Meetings::PresentationMode::ShowComponent.new(
     meeting: @meeting,
     sorted_agenda_item_ids:,
     current_id: @current_id,
